Refactor project organization JC-627
