+++ apps/vanilla/src/pages/blackjack/ts/core/turnGenerator.ts
