+++ apps/vanilla/src/pages/blackjack/ts/utils/subscriber.ts
